pyproject.toml
[project]
name = "ragtemplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.39.4",
    "decorator>=5.2.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "imageio[ffmpeg]>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "moviepy>=2.2.1",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "opencv-python-headless>=4.12.0.88",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "tqdm>=4.67.1",
    "livekit>=1.0.11",
    "livekit-agents>=1.1.6",
    "livekit-plugins-sarvam>=1.1.6",
    "livekit-plugins-silero>=1.1.6",
]
